MiniCalculatorV3
version 3 mini calculator gets input(integer) as well as the arithmatic operators such as +, -, * ,/ %(mod) from the user and operates on them.
if the arithmatic operator entered by the user is invalid, the program will produce an error message and ask the user to enter the proper arithmatic operator. The program will keep asking the user for the proper arithmatic operator untill the right operator is entered. The program also asks the user if s/he is done or wants to calculate more. the program will end when the user answers no.